Easy prod/debug swtich per network
