+++ chrome/manifest.staging.js
@@ -10,11 +10,11 @@ import {
   genCSP,
 } from './constants';
 
-export default buildManifest({
+export default (isDebug: boolean) => buildManifest({
   description: '[staging] Cardano ADA wallet',
   defaultTitle: '[staging] Yoroi',
   contentSecurityPolicy: genCSP({
-    isDev: false,
+    isDev: isDebug,
     additional: {
       'connect-src': [serverToPermission(Servers.ByronStaging)],
     },
