+++ package.json
@@ -3,11 +3,12 @@
   "version": "1.10.0",
   "description": "Cardano ADA wallet",
   "scripts": {
-    "dev": "babel-node scripts/start --env 'development'",
-    "shelley-dev": "babel-node scripts/start --env 'shelley-dev'",
-    "shelley-testnet": "babel-node scripts/build --env 'shelley-testnet'",
+    "byron-dev": "babel-node scripts/build --type=debug --env 'mainnet'",
+    "byron-prod": "babel-node scripts/build --type=prod --env 'mainnet'",
+    "shelley-dev": "babel-node scripts/build --type=debug --env 'shelley-testnet'",
+    "shelley-prod": "babel-node scripts/build --type=prod --env 'shelley-testnet'",
     "dev-mock-backend": "babel-node scripts/startWithMockServer.js",
-    "build": "babel-node scripts/build",
+    "build": "babel-node scripts/build --type=prod",
     "compress": "babel-node scripts/compress",
     "compress-e2e-tests": "npm run compress -- --env 'test' --app-id 'yoroi' --codebase 'https://yoroiwallet.com/dw/yoroi-test-extension.crx' --key ./e2etest-key.pem",
     "build-compress:shelley": "npm run build -- --env 'shelley-testnet' && npm run compress -- --env 'shelley-testnet' --app-id 'yoroi' --zip-only --codebase 'https://yoroiwallet.com/dw/yoroi-extension.crx' --key ./shelley-production.pem",
