+++ scripts/build.js
@@ -1,22 +1,62 @@
 // @flow
+
 const tasks = require('./tasks');
 const argv = require('minimist')(process.argv.slice(2));
-const shell = require('shelljs');
 
-// ovrerride NODE_ENV for ConfigWebpackPlugin
+// override NODE_ENV for ConfigWebpackPlugin
 process.env.NODE_CONFIG_ENV = argv.env;
 
-tasks.replaceWebpack();
 
-console.log('[Build manifest]');
-console.log('-'.repeat(80));
-tasks.buildManifests();
+export function buildProd(env: string) {
+  const shell = require('shelljs');
 
-console.log('[Copy assets]');
-console.log('-'.repeat(80));
-tasks.copyAssets('build', argv.env);
+  tasks.replaceWebpack();
 
-console.log('[Webpack Build]');
-console.log('-'.repeat(80));
+  console.log('[Build manifest]');
+  console.log('-'.repeat(80));
+  tasks.buildManifests(false);
 
-process.exit(shell.exec(`./node_modules/.bin/webpack --config webpack/prodConfig.js --progress --profile --colors --env=${argv.env}`).code);
+  console.log('[Copy assets]');
+  console.log('-'.repeat(80));
+  tasks.copyAssets('build', env);
+
+  console.log('[Webpack Build]');
+  console.log('-'.repeat(80));
+
+  process.exit(shell.exec(`./node_modules/.bin/webpack --config webpack/prodConfig.js --progress --profile --colors --env=${argv.env}`).code);
+}
+
+export function buildDev(env: string) {
+  const connections = require('./connections');
+  const createWebpackServer = require('webpack-httpolyglot-server');
+
+  const config = require(`../webpack/devConfig`);
+
+  tasks.replaceWebpack();
+
+  console.log('[Build manifest]');
+  console.log('-'.repeat(80));
+  tasks.buildManifests(true);
+
+  console.log('[Copy assets]');
+  console.log('-'.repeat(80));
+  tasks.copyAssets('dev', env);
+
+  console.log('[Webpack Dev]');
+  console.log('-'.repeat(80));
+  console.log('If you\'re developing Inject page,');
+  console.log(`please allow 'https://localhost:${connections.Ports.WebpackDev}' connections in Google Chrome,`);
+  console.log('and load unpacked extensions with `./dev` folder. (see https://developer.chrome.com/extensions/getstarted#unpacked)\n');
+  createWebpackServer(config.baseDevConfig(argv.env), {
+    host: 'localhost',
+    port: connections.Ports.WebpackDev
+  });
+}
+
+if (argv.type === 'debug') {
+  buildDev(argv.env);
+} else if (argv.type === 'prod') {
+  buildProd(argv.env);
+} else {
+  throw new Error(`Unknown type ${argv.type}`);
+}
